refactor(js): retire app/javascript/app/ directory
