--- app/javascript/app/index.js
@@ -1,2 +1,0 @@
-import "./remote_form_render";
-import "./live_leaderboard";
