+++ app/javascript/lib/live_leaderboard.js
